pyproject.toml
[tool.poetry]
name = "django-rest-captcha"
version = "0.3.0"
description = "Lightweight version of django-simple-captcha for work with django-rest-framework"
readme = "README.md"
authors = ["evgeny.zuev <zueves@gmail.com>", "e9000000000 <skipper224483@gmail.com>"]
license = "MIT"
packages = [
    {include = "rest_captcha"},
]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "3.13.1"
django="^4.0.4"
Pillow = "9.0.0"

[tool.poetry.dev-dependencies]
mypy = "^0.950"
black = "^22.3.0"
pylama = "^8.3.8"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
tests = "runtests:main"
